feat: ensure cart product quantity can be decreased

+++ src/actions/decrease-cart-product-quantity/index.ts
@@ -8,12 +8,12 @@ import { db } from "@/db";
 import { cartItemTable } from "@/db/schema";
 import { auth } from "@/lib/auth";
 
-import { removeProductFromCartSchema } from "./schema";
+import { decreaseCartProductQuantitySchema } from "./schema";
 
-export const removeProductFromCart = async (
-  data: z.infer<typeof removeProductFromCartSchema>,
+export const decreaseCartProductQuantity = async (
+  data: z.infer<typeof decreaseCartProductQuantitySchema>,
 ) => {
-  removeProductFromCartSchema.parse(data);
+  decreaseCartProductQuantitySchema.parse(data);
   const session = await auth.api.getSession({
     headers: await headers(),
   });
@@ -28,7 +28,6 @@ export const removeProductFromCart = async (
       cart: true,
     },
   });
-
   if (!cartItem) {
     throw new Error("Product variant not found in cart");
   }
@@ -37,5 +36,12 @@ export const removeProductFromCart = async (
     throw new Error("Uauthorized");
   }
 
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+  if (cartItem.quantity === 1) {
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+  }
+
+  await db
+    .update(cartItemTable)
+    .set({ quantity: cartItem.quantity - 1 })
+    .where(eq(cartItemTable.id, cartItem.id));
 };
